+++ P-600-G/P-600-GUI 2026/Project_Source_Files/helpers/H_00-P_Build_Tree.h
@@ -13,9 +13,9 @@ struct Build_Tree :
 	public Base_Build_Tree
 {
 	static ValueTree exposed_parameter(Identifier id, String name, Ctrl_Type ctrl, 
-									   uint8 first_nybble_index, uint8 first_bit_index,
-									   uint8 bit_count, uint8 choice_count, 
-									   uint8 init_choice,int ctrl_center_x, int ctrl_center_y, 
+									   int first_nybble_index, int first_bit_index,
+									   int bit_count, int choice_count, 
+									   int init_choice,int ctrl_center_x, int ctrl_center_y, 
 									   int ctrl_w, int ctrl_h, String description,
 									   ValueTree curt_choice_names, ValueTree choice_names);
